+++ packages/zent-button/__tests__/Button.js
@@ -151,7 +151,7 @@ describe('<Button />', () => {
     expect(isClicked).toBe(false);
   });
 
-  test('HtmlType Button', () => {
+  test('Button htmlType', () => {
     mount(<Button htmlType="submit" />);
     expect(buttonNode.type).toBe('submit');
 
@@ -166,4 +166,21 @@ describe('<Button />', () => {
     mount(<Button style={{ fontSize: '20px' }} />);
     expect(buttonNode.style.fontSize).toBe('20px');
   });
+
+  test('Component', () => {
+    function Link({ to, children, ...rest }) {
+      return <a href={`/#${to}`} {...rest}>{children}</a>;
+    }
+    mount(<Button to="/path" component={Link} />);
+
+    expect(buttonNode.href).toBe('/#/path');
+    expect(buttonNode.classList.contains('zent-btn')).toBe(true);
+    expect(buttonNode.tagName.toLowerCase()).toBe('a');
+  });
+
+  test('Link with additional props', () => {
+    mount(<Button href="http://www.youzan.com" download="foobar" />);
+
+    expect(buttonNode.download).toBe('foobar');
+  })
 });